+++ Gruntfile.js
@@ -15,33 +15,6 @@ module.exports = function (grunt) {
   // Project configuration.
   grunt.initConfig({
 
-    sed: {
-      h2: {
-        path: 'README.txt',
-        pattern: /^##([^#]+)#*?\s*?\n/gm,
-        replacement: '==$1==\n',
-        recursive: true
-      },
-      del_sc: {
-        path: 'README.txt',
-        pattern: /^\[(youtube|vimeo)[^\]]+\]/gm,
-        replacement: '',
-        recursive: true
-      },
-      sc_start: {
-        path: 'README.txt',
-        pattern: /`\[\[/gm,
-        replacement: '`[',
-        recursive: true
-      },
-      sc_end: {
-        path: 'README.txt',
-        pattern: /\]\]`/gm,
-        replacement: ']`',
-        recursive: true
-      }
-    },
-
     concat: {
       options: {
         separator: '\n\n',
@@ -49,15 +22,12 @@ module.exports = function (grunt) {
       readme_md: {
         src: [
            'readme/description.md',
-           'readme/description-intro.md',
-           'readme/description-links.md',
-           'readme/description-features.md',
+           'readme/description-lead.html',
+           'readme/description-features.html',
+           'readme/description-features-pro.html',
            'readme/description-supported-providers.md',
-           'readme/description-quick-intro.md',
-           'readme/description-pro-addon-demo.md',
-           'readme/description-hd-thumbnail.md',
-           'readme/description-reviews.md',
-           'readme/description-roadmap.md',
+           'readme/description-only-wp-org.md',
+           'readme/description-reviews.html',
            'readme/installation.md',
            'readme/faq.md',
            'readme/screenshots.md',
@@ -79,5 +49,5 @@ module.exports = function (grunt) {
   require('load-grunt-tasks')(grunt, { scope: 'devDependencies' });
 
   // Default task.
-  grunt.registerTask('default', ['concat', 'sed']);
+  grunt.registerTask('default', ['concat']);
 };